// core/models/Inventario.ts
export interface Inventario {
  id: number;
  nombre: string;
  aplicaVentas: boolean;
  aplicaCompras: boolean;
  suspendido: boolean;
  otrosF1: string;
  equipo: string;
  otrosN1: number;
  otrosN2: number;
  otrosC1: null;
  otrosC2: null;
  otrosC3: null;
  otrosC4: null;
  otrosT1: null;
  fechaRegistro: string;
  usuarioRegistroNombre: string;
  fechaModificacion: string;
  usuarioModificacionNombre: string;
}